requirements.txt
Flask>=2.3.0
numpy>=1.26.0
scipy>=1.11.0
pyserial>=3.5
keyboard>=0.13.5
matplotlib>=3.8.0
pandas>=2.1.0
scikit-learn>=1.3.0
joblib~=1.4.2
filterpy~=1.4.5